requirements.txt
fastapi
uvicorn
web3
python-dotenv
ipfshttpclient
pydantic
eth_account
